requirements.txt
fastapi==0.116.1
uvicorn[standard]==0.35.0
pydantic-settings==2.10.1
SQLAlchemy==2.0.43
alembic==1.16.4
psycopg[binary]==3.2.9
greenlet==3.2.4
PyJWT==2.10.1
redis==6.4.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.20
email-validator==2.2.0
celery[redis]==5.5.3
pytest==8.4.1
pytest-asyncio==1.1.0
pytest-cov==6.2.1
httpx==0.28.1
ruff==0.12.8
